refactor: extract abstract class
