+++ packages/openapi-code-generator/src/typescript/typescript-koa/schema-builders/joi-schema-builder.ts
@@ -1,8 +1,8 @@
-import { Input } from "../../../core/input"
-import { IRModelObject, IRParameter, MaybeIRModel } from "../../../core/openapi-types-normalized"
-import { isDefined } from "../../../core/utils"
-import { SchemaBuilder } from "./schema-builder"
-import { ImportBuilder } from "../../common/import-builder"
+import {Input} from "../../../core/input"
+import {IRModelString} from "../../../core/openapi-types-normalized"
+import {isDefined} from "../../../core/utils"
+import {SchemaBuilder} from "./schema-builder"
+import {ImportBuilder} from "../../common/import-builder"
 
 enum JoiFn {
   Object = "object()",
@@ -13,85 +13,38 @@ enum JoiFn {
   Required = "required()"
 }
 
-export class JoiBuilder implements SchemaBuilder {
+export class JoiBuilder extends SchemaBuilder {
   constructor(
     private readonly joi = "joi",
-    private readonly input: Input,
+    input: Input,
   ) {
+    super(input)
   }
-
 
   importHelpers(importBuilder: ImportBuilder) {
     importBuilder.from("@nahkies/typescript-koa-runtime/joi")
       .add("queryValidationFactory", "paramValidationFactory", "bodyValidationFactory")
   }
 
-  fromParameters(parameters: IRParameter[]): string {
-    const model: IRModelObject = {
-      type: "object",
-      required: [],
-      properties: {},
-      allOf: [],
-      readOnly: false,
-      nullable: false,
-      additionalProperties: false,
-    }
-
-    parameters.forEach(parameter => {
-      if (parameter.required) {
-        model.required.push(parameter.name)
-      }
-      model.properties[parameter.name] = parameter.schema
-    })
-
-    return this.fromModel(model, true)
-  }
-
-  fromModel(maybeModel: MaybeIRModel, required: boolean): string {
-    const model = this.input.schema(maybeModel)
-
-    switch (model.type) {
-      case "string":
-        return this.string(required)
-      case "number":
-        return this.number(required)
-      case "boolean":
-        return this.boolean(required)
-      case "array":
-        return this.array([
-          this.fromModel(this.input.schema(model.items), false),
-        ], required)
-      case "object":
-        return this.object(
-          Object.fromEntries(
-            Object.entries(model.properties)
-              .map(([key, value]) => {
-                return [key, this.fromModel(this.input.schema(value), model.required.includes(key))]
-              }),
-          ), required,
-        )
-    }
-  }
-
-  private object(keys: Record<string, string>, required: boolean): string {
+  protected object(keys: Record<string, string>, required: boolean): string {
     return [
       this.joi,
       JoiFn.Object,
-      `keys({${ Object.entries(keys).map(([key, value]) => `"${ key }": ${ value }`).join(",") } })`,
+      `keys({${Object.entries(keys).map(([key, value]) => `"${key}": ${value}`).join(",")} })`,
       required ? JoiFn.Required : undefined,
     ].filter(isDefined).join(".")
   }
 
-  private array(items: string[], required: boolean): string {
+  protected array(items: string[], required: boolean): string {
     return [
       this.joi,
       JoiFn.Array,
-      `items(${ items.join(",") })`,
+      `items(${items.join(",")})`,
       required ? JoiFn.Required : undefined,
     ].filter(isDefined).join(".")
   }
 
-  private number(required: boolean) {
+  protected number(required: boolean) {
     return [
       this.joi,
       JoiFn.Number,
@@ -99,7 +52,9 @@ export class JoiBuilder implements SchemaBuilder {
     ].filter(isDefined).join(".")
   }
 
-  private string(required: boolean) {
+  protected string(model: IRModelString, required: boolean) {
+    // TODO: enum support
+
     return [
       this.joi,
       JoiFn.String,
@@ -107,7 +62,7 @@ export class JoiBuilder implements SchemaBuilder {
     ].filter(isDefined).join(".")
   }
 
-  private boolean(required: boolean) {
+  protected boolean(required: boolean) {
     return [
       this.joi,
       JoiFn.Boolean,
